chore(test): split the review-scratch trap tests into their own file
test_review_scratch.py crossed the 300-line bloat guideline after the
round-4 trap-exit-status tests were added. Split those three integration
tests (plus their local bash() resolver and _run_trap_pattern helper) out
into test_review_scratch_trap_exit_status.py — a distinct concern (the
shell trap idiom used by the skill docs) from the unit-level
resolve()/cleanup() coverage the original file keeps.

Run-ID: iterate-2026-09-03-review-scratch-path
